kite_capture/requirements.txt
kiteconnect==5.0.1
pyarrow>=14.0.0
polars>=0.20.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
zstandard>=0.22.0
